aarch64 aspec: use map union for ptes_of levels
Use a more principled way to define ptes_of/get_pte by defining
level_pte_of parametric in the level and setting ptes_of to the union
of all levels. This works because objects must be distinct.

For store_pte a simple union doesn't work, but we can still first
extract the level, and then use the level to update the object for that
level.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>